+++ example_datagram_transfer.py
@@ -13,7 +13,6 @@ import threading
 
 from openlcb.tcpsocket import TcpSocket
 from canbus.canphysicallayergridconnect import CanPhysicalLayerGridConnect
-# from canbus.canframe import CanFrame
 from canbus.canlink import CanLink
 # from openlcb.controlframe import ControlFrame
 from openlcb.nodeid import NodeID
